+++ srcs/ftlib/ft_freturn.c
@@ -6,13 +6,13 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/04 01:52:23 by abaurens          #+#    #+#             */
-/*   Updated: 2018/12/13 18:48:32 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/03/30 15:51:59 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-long long int	ft_freturn(void *ptr, long long int ret)
+long int	ft_freturn(void *ptr, long int ret)
 {
 	free(ptr);
 	return (ret);
