+++ srcs/ftlib/ft_max.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_maximum.c                                       :+:      :+:    :+:   */
+/*   ft_max.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/13 16:51:24 by abaurens          #+#    #+#             */
-/*   Updated: 2018/12/13 17:03:57 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/03/30 15:48:01 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long long int	ft_max(long long int a, long long int b)
+long int	ft_max(long int a, long int b)
 {
 	return (a > b ? a : b);
 }
