+++ srcs/ftlib/ft_numlen.c
@@ -6,13 +6,13 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/12 23:25:16 by abaurens          #+#    #+#             */
-/*   Updated: 2018/12/13 17:12:52 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/03/27 17:28:52 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <string.h>
 
-size_t		ft_numlen(long long int nb)
+size_t		ft_numlen(long int nb)
 {
 	size_t	len;
